Agregado de loading skeletons para las paginas de carga de datos

+++ client/src/components/Producto.jsx
@@ -9,7 +9,7 @@ import { useNavigate } from "react-router-dom";
 import { useState } from "react";
 
 // Importo iconos de React Icons
-import { MdDelete, MdFileDownloadDone } from 'react-icons/md';
+import { MdDelete, MdOutlineRestoreFromTrash } from 'react-icons/md';
 import { AiFillEdit } from 'react-icons/ai';
 
 // Ruta de la imagen por defecto
@@ -17,36 +17,41 @@ import defaultImage from "../../assets/images/NoImage.jpg";
 
 // El componente recibe un elemento del arreglo de productos
 // y muestra todas las propiedades del mismo
+
 export default function ProductCard({ product }) {
+    // Extraigo del context las funciones para eliminar y restaurar un producto
+    const { deleteProduct, restoreProduct } = useGlobalContext();
 
-    // Extraigo del context las funciones para eliminar un producto
-    const {deleteProduct} = useGlobalContext();
-
-    // Declaro constante para disponer del useNavigate
+     // Declaro constante para disponer del useNavigate
     const navigate = useNavigate();
+
+    // Estado para mostrar la confirmacion de eliminacion o restauracion
+    const [showDeleteOrRestoreConfirmation, setShowDeleteOrRestoreConfirmation] = useState(false);
 
     // Truncar el titulo si es que es demasiado largo
     const maxLength = 23;
     let title;
     product.productName.length <= maxLength ? title = product.productName : (title = product.productName.slice(0, maxLength) + '...');
 
-    
-    // CONFIRMACION DE ELIMINACION
-    const [showDeleteConfirmation, setShowDeleteConfirmation] = useState(false);
-    // Funcion para cambiar el estado de la variable que muestra el modal de confirmacion de eliminacion
-    const handleDelete = () => {
-        setShowDeleteConfirmation(true);
-    }
-    // CONFIRMAR: Se llama a la funcion que elimina el producto y se restablece showDeleteConfirmation a false
-    const handleDeleteConfirm = async () => {
-        // Llamar a la funcion que se encarga de eliminar el producto segun su id
-        deleteProduct(product.id);
-        // Restablece el estado de showDeleteConfirmation
-        setShowDeleteConfirmation(false);
+    // Handler para la confirmacion
+    const handleDeleteOrRestore = () => {
+        setShowDeleteOrRestoreConfirmation(true);
     };
-    // CANCELAR: Se cambia a false el valor de showDelteConfirmation para ocultar el modal
-    const handleDeleteCancel = () => {
-        setShowDeleteConfirmation(false);
+
+    // Confirmar
+    const handleDeleteOrRestoreConfirm = async () => {
+        setShowDeleteOrRestoreConfirmation(false);
+        if (product.active) {
+            deleteProduct(product.id);
+        } else {
+            restoreProduct(product.id);
+            window.location.reload(); // Recargar la página después de restaurar el producto
+        }
+    };
+
+    // Cancelar
+    const handleDeleteOrRestoreCancel = () => {
+        setShowDeleteOrRestoreConfirmation(false);
     };
 
     return (
@@ -63,23 +68,30 @@ export default function ProductCard({ product }) {
             <p className='Tarjeta_field mb-10'>Precio: ${product.price}</p>
             {/* <p className='Tarjeta_field'><span>C: {product.createdAt}</span></p> */}
             {/* <p className='Tarjeta_field'><span>M: {product.updatedAt}</span></p> */}
+
             <footer className='absolute bottom-3 flex flex-wrap gap-x-2'>
-                <button className='Card-icon' onClick={() => handleDelete()}><MdDelete/></button>
+                <button className='Card-icon' onClick={() => handleDeleteOrRestore()}>{product.active ? <MdDelete/> : <MdOutlineRestoreFromTrash/>}</button>
                 <button className='Card-icon' onClick={() => navigate(`/editarProducto/${product.id}`)}><AiFillEdit/></button>
             </footer>
 
-            {showDeleteConfirmation && (
-            <div>
-                <div className='backdrop-blur-md p-3 absolute inset-0 w-full h-full'></div>
-                <div className='Tarjeta flex flex-wrap w-10/12 p-2 absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2'>
-                    <p className='w-full text-center'>¿Estas seguro que deseas eliminar este producto?</p>
-                    <div className='w-full flex justify-between'>
-                        <button type='buton' className = 'MainButton w-24 mt-2' onClick={handleDeleteConfirm}>Confirmar</button>
-                        <button type='buton' className = 'MainButton w-24 mt-2' onClick={handleDeleteCancel}>Cancelar</button>
+            {showDeleteOrRestoreConfirmation && (
+                <div>
+                    <div className='backdrop-blur-md p-3 absolute inset-0 w-full h-full'></div>
+                    <div className='Tarjeta flex flex-wrap w-10/12 p-2 absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2'>
+                        <p className='w-full text-center'>
+                            {product.active ? '¿Estás seguro que deseas eliminar este producto?' : '¿Estás seguro que deseas restaurar este producto?'}
+                        </p>
+                        <div className='w-full flex justify-between'>
+                            <button type='button' className='MainButton w-24 mt-2' onClick={handleDeleteOrRestoreConfirm}>
+                                {product.active ? 'Confirmar' : 'Restaurar'}
+                            </button>
+                            <button type='button' className='MainButton w-24 mt-2' onClick={handleDeleteOrRestoreCancel}>
+                                Cancelar
+                            </button>
+                        </div>
                     </div>
                 </div>
-            </div>
             )}
         </div>
-  )
+    );
 }
