+++ server/controllers/products.controllers.js
@@ -76,6 +76,10 @@ export const updateProduct = async (req, res) => {
     let result = null;
     
     if (req.files && req.files.imgURL) {
+        // Si se envia una imagen en el req
+        // console.log("IMAGEN ENVIADA")
+
+        // Se sube la imagen
         const resultUpload = await uploadImage(req.files.imgURL.tempFilePath)
         // Una vez se subio la imagen, se la eliminara de la carpeta temporal
         await fs.remove(req.files.imgURL.tempFilePath);
@@ -99,20 +103,26 @@ export const updateProduct = async (req, res) => {
         // Ejecutar el UPDATE incluyendo la nueva imagen del producto
         result = await pool.query("UPDATE products SET productName = ?, imgURL = ?, imgPublic_id = ?, description = ?, price = ? WHERE id = ?", [productName, image.url, image.public_id, description, price, req.params.id]);
     } else {
-        // En este caso, no se incluyó una nueva imagen en el req, por lo que se verificara si el producto ya contaba con una imagen
-        const [product] = await pool.query("SELECT * FROM products WHERE id = ?", [req.params.id]);
-        let cloudinaryImgPublic_id = product[0].imgPublic_id; // Si no tenia imagen, la variable tomará el valor de null
-    
-        if (cloudinaryImgPublic_id) {
-          // Si tenia imagen, se elimina de Cloudinary y se actualiza el registro en la base de datos
-          await deleteImage(cloudinaryImgPublic_id);
-    
-          // Se actualiza el producto limpiando el contenido anterior de la imagen (imgURL e imgPublic_id )
-          result = await pool.query("UPDATE products SET productName = ?, description = ?, price = ?, imgURL = NULL, imgPublic_id = NULL WHERE id = ?", [productName, description, price, req.params.id]);
-        } else {
-          // Si no había una imagen anterior, simplemente actualizamos el registro en la base de datos excluyendo el campo imgURL.
-          result = await pool.query("UPDATE products SET productName = ?, description = ?, price = ? WHERE id = ?", [productName, description, price, req.params.id]);
-        }
+        // Si NO se envia una imagen en el req
+        // console.log("IMAGEN NO ENVIADA");
+
+        // Si se desea quitar la imagen del producto
+        if (req.body.imgURL === "DELETE_IMAGE") {
+            // Eliminar la imagen del producto
+            const [product] = await pool.query("SELECT * FROM products WHERE id = ?", [req.params.id]);
+            let cloudinaryImgPublic_id = product[0].imgPublic_id;
+            
+            if (cloudinaryImgPublic_id) {
+              await deleteImage(cloudinaryImgPublic_id);
+            }
+            
+            // Actualizar el producto eliminando la imagen
+            result = await pool.query("UPDATE products SET productName = ?, description = ?, price = ?, imgURL = NULL, imgPublic_id = NULL WHERE id = ?", [productName, description, price, req.params.id]);
+          } else {
+            // Si no se envía una nueva imagen ni una solicitud de eliminacion de la misma, se mantiene la imagen que ya tenia 
+            // Se ejecuta el update sin modificar los campos que contiene la informacion de la imagen
+            result = await pool.query("UPDATE products SET productName = ?, description = ?, price = ? WHERE id = ?", [productName, description, price, req.params.id]);
+          }
       }
     
     /* 
@@ -127,11 +137,11 @@ export const updateProduct = async (req, res) => {
     return res.sendStatus(204);
 }
 
-// * Funcion para obtener todos los productos de la db
+// Funcion para obtener todos los productos de la db
 export const getProducts = async (req, res) => {
     try {
         //res.send('Obteniendo productos')
-        const [result] = await pool.query("SELECT * FROM products ORDER BY createdAt DESC");
+        const [result] = await pool.query("SELECT * FROM products WHERE active = true ORDER BY createdAt DESC");
         // Devuelve un arreglo con todos los productos
         //console.log(result);
         //console.log("Obteniendo productos...")
@@ -140,10 +150,19 @@ export const getProducts = async (req, res) => {
         // Retornar el mensaje de error como respuesta
         return res.status(500).json({message: error.message});
     }
-    
 }
 
-// * Funcion para obtener un producto mediante su id
+// Función para obtener solo los productos inactivos de la base de datos
+export const getInactiveProducts = async (req, res) => {
+    try {
+        const [result] = await pool.query("SELECT * FROM products WHERE active = false ORDER BY createdAt DESC");
+        res.json(result); // Devuelve un arreglo con los productos inactivos
+    } catch (error) {
+        return res.status(500).json({ message: error.message });
+    }
+}
+
+// Funcion para obtener un producto mediante su id
 export const getProduct = async (req, res) => {
     try {
         //res.send('Obteniendo un producto')
@@ -166,35 +185,60 @@ export const getProduct = async (req, res) => {
 // Funcion para eliminar un producto mediante su id
 export const deleteProduct = async (req, res) => {
     try {
-        //res.send('Eliminando un producto')
-
-        // Primero se debe encontrar el imgPublic_id de cloudinary que tiene la imagen del producto
+        // Verificar si existe el producto a eliminar
         const [product] = await pool.query("SELECT * FROM products WHERE id = ?", [req.params.id]);
-        let cloudinaryImgPublic_id = product[0].imgPublic_id; // En caso de no tener imagen, la variable tomara el valor de null
+        if (product.length === 0) {
+            return res.status(404).json({ message: "Producto no encontrado" });
+        }
 
-        /* 
-        Aca, si bien no es necesario devolver un resultado luego de eliminar un producto,
-        Se utilizara el result.affectedRows para saber si el producto que se desea eliminar
-        realmente existe en la db, en el caso de no hacerlo se devolvera un msg de error
-        */
-        const [result] = await pool.query("DELETE FROM products WHERE id = ?", [req.params.id]);
-        if (result.affectedRows === 0) {
-            // NO EXISTE EL PRODUCTO
-            return res.status(404).json({message: "Producto no encontrado"});
+        // Definir el id del producto y si esta asociado a algun pedido
+        const productId = product[0].id;
+        const isProductInOrders = await checkProductInOrders(productId);
+
+        // Determinar si sera un borrado logico o fisico 
+        if (isProductInOrders) {
+            // Producto asociado a pedidos, cambiar active a false
+            await pool.query("UPDATE products SET active = false WHERE id = ?", [productId]);
+            console.log('Producto desactivado');
+            return res.sendStatus(204); // Producto desactivado con éxito
         } else {
-            // SI EXISTE EL PRODUCTO
-            // Con el public_id de cloudinary se ejecuta la funcion para eliminar la imagen en el caso de tener
+            // No asociado a pedidos, eliminar físicamente
+            const cloudinaryImgPublic_id = product[0].imgPublic_id;
+            await pool.query("DELETE FROM products WHERE id = ?", [productId]);
+
             if (cloudinaryImgPublic_id) {
                 await deleteImage(cloudinaryImgPublic_id);
             }
-            console.log('Producto Eliminado');
-            // Finalmente el estado sera 204 indicando que se elimino correctamente pero no devuelve ningun resultado
-            return res.sendStatus(204);
-        }
-        
+
+            console.log('Producto eliminado');
+            return res.sendStatus(204); // Producto eliminado con éxito
+        }  
     } catch (error) {
-        // Retornar el mensaje de error como respuesta
-        return res.status(500).json({message: error.message});
+        return res.status(500).json({ message: error.message });
+    }
+}
+
+// Funcion para checkear si un producto esta asociado a uno o mas pedidos
+async function checkProductInOrders(productId) {
+    const [ordersCount] = await pool.query("SELECT COUNT(*) AS count FROM orders_products WHERE product_id = ?", [productId]);
+    return ordersCount[0].count > 0;
+}
+
+// Función para restaurar un producto
+export const restoreProduct = async (req, res) => {
+    try {
+        // Verificar si existe el producto a restaurar
+        const [product] = await pool.query("SELECT * FROM products WHERE id = ?", [req.params.id]);
+        if (product.length === 0) {
+            return res.status(404).json({ message: "Producto no encontrado" });
+        }
+
+        // Restaurar el producto (cambiar active a true)
+        await pool.query("UPDATE products SET active = true WHERE id = ?", [req.params.id]);
+        console.log('Producto restaurado');
+        return res.sendStatus(204); // Producto restaurado con éxito
+    } catch (error) {
+        return res.status(500).json({ message: error.message });
     }
 }
